requirements.txt
# Core
Flask==3.1.3
python-dotenv==1.2.2

# LLM Framework
langchain==1.2.15
langchain-core==1.3.0

# LLM Providers (agent auto-detects which key is set)
langchain-google-genai==4.2.2
langchain-groq==1.1.2
